+++ src/draw_context.py
@@ -18,32 +18,32 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-import math
+import logging
 
 from gi.repository import Pango, Adw
+
+logger = logging.getLogger(__name__)
 
 
 class DrawContext:
     """Store the parameters required for drawing the ruler."""
 
-    def __init__(self, settings) -> None:
+    def __init__(self, settings, monitors) -> None:
         """Initialize the object.
 
         :param settings: Application settings.
         :type settings: :py:class:``GSettings``
+        :param monitors: Application settings.
+        :type monitors: :py:class:``monitor_mngt.MonitorMngt``
         """
-
         self.settings = settings
-        self._warning_dialog = None
+        self.monitors = monitors
 
-        # Gdk.Monitor
-        self._monitor = None
-        # Monitor width
-        self.monitor_width: int = 0
-        # Monitor height
-        self.monitor_height: int = 0
-        # Pixels per millimeter
-        self._ppmm: float = 0.0
+        # Warning dialog window per monitor
+        self._warning_dialogs = {}
+
+        # Monitor currently used by the ruler (monitor_mngt.Monitor)
+        self.current_monitor = None
 
         # Cairo context
         self.ctx = None
@@ -85,23 +85,25 @@ class DrawContext:
         # Scale direction
         self.left2right: bool = True
 
-        # Monitor size
-        self.compute_monitor_size: bool = True
-        self.monitor_diagonal_inch: float = 24.0
-
     def refresh_from_settings(self) -> None:
         """Reload the parameter from GSettings."""
-
         self.track_pointer = self.settings.get_boolean("track-pointer")
         self.offset = self.settings.get_double("offset")
+
+        logger.debug(f"Get settings:     track-pointer: {self.track_pointer}")
+        logger.debug(f"Get settings:            offset: {self.offset}")
 
         # Colors
         if self.settings.get_boolean("use-default-color"):
             self.color_fg = [0.0, 0.0, 0.0, 1.0]
             self.color_bg = [1.0, 1.0, 1.0, 1.0]
+            logger.debug("Get settings: use-default-color: True")
         else:
             self.color_fg = self.settings.get_value("foreground-color")
             self.color_bg = self.settings.get_value("background-color")
+            logger.debug("Get settings: use-default-color: False")
+            logger.debug(f"Get settings:  foreground-color: {self.color_fg}")
+            logger.debug(f"Get settings:  background-color: {self.color_bg}")
 
         # Font
         font = (
@@ -109,6 +111,7 @@ class DrawContext:
             if self.settings.get_boolean("use-default-font")
             else self.settings.get_string("font-name")
         )
+        logger.debug(f"Get settings:              font: {font}")
         if font != self.font_name:
             self.font_name = font
             self.font_desc = Pango.font_description_from_string(font)
@@ -118,99 +121,76 @@ class DrawContext:
 
         # Scale direction
         self.left2right = self.settings.get_boolean("direction-left-to-right")
+        logger.debug(f"Get settings: direction-left-to-right: {self.left2right}")
 
-        # Monitor size
-        self.compute_monitor_size = self.settings.get_boolean("compute-monitor-size")
-        self.monitor_diagonal_inch = self.settings.get_double("monitor-size")
+        # Monitor calibration parameters
+        self.monitors.get_settings()
 
     def warning(self) -> None:
         """Display a warning message when the system does not provide the monitor size.
 
-        The dialog is only displayed once.
+        The dialog is only displayed once per monitor.
         """
-
-        if not self._warning_dialog:
-            self._warning_dialog = Adw.AlertDialog()
-            self._warning_dialog.set_heading(_("Monitor Size Unknown"))
-            self._warning_dialog.set_body(
+        monitor_name = self.current_monitor.name
+        if not self._warning_dialogs.get(monitor_name):
+            dialog = Adw.AlertDialog()
+            dialog.set_heading(_("Monitor Size Unknown"))
+            dialog.set_body(
                 _(
-                    "The monitor size cannot be determined.\n"
+                    f"The size of the {monitor_name} monitor cannot be determined.\n"
                     "Use the Preferences dialog to calibrate Length for "
                     "your monitor size.\n"
                     "In the meantime, computations are done for a "
-                    f"{round(self.monitor_diagonal_inch, 1)} "
+                    f"{round(self.current_monitor.diag_inch, 1)} "
                     "inches monitor."
                 )
             )
-            self._warning_dialog.add_response("ok", _("OK"))
-            self._warning_dialog.set_default_response("ok")
-            self._warning_dialog.set_close_response("ok")
-            self._warning_dialog.present()
+            dialog.add_response("ok", _("OK"))
+            dialog.set_default_response("ok")
+            dialog.set_close_response("ok")
+            dialog.present()
+            self._warning_dialogs[monitor_name] = dialog
 
     def set_monitor(self, monitor) -> None:
         """Set the monitor object that is used to compute some ruler units.
 
         :param monitor: The object used to convert the unit into pixels.
-        :type monitor: :py:class:``Gdk.Monitor``
+        :type monitor: :py:class:``monitor_mngt.Monitor``
         """
-        self._monitor = monitor
-        if monitor:
-            geometry = monitor.get_geometry()
-            self.monitor_width = geometry.width
-            self.monitor_height = geometry.height
-        else:
-            # monitor should not be None
-            self.monitor_width = 1920
-            self.monitor_height = 1080
-        self._ppmm = 0.0
-
-    def _compute_ppmm(self) -> float:
-        """Compute and return the number of pixels per millimeter (ppmm).
-
-        The computation uses the user provided monitor diagonal size (in the
-        ``monitor-size`` GSettings parameter).
-        """
-        if self._ppmm:
-            return self._ppmm
-        ppi = self.monitor_width / math.sqrt(
-            (self.monitor_diagonal_inch**2)
-            / (1 + (self.monitor_height / self.monitor_width) ** 2)
-        )
-        self._ppmm = ppi / 25.4  # Convert from inch to millimeter
-        return self._ppmm
+        self.current_monitor = monitor
 
     @property
     def ppmm_width(self) -> float:
         """Return the number of pixels per millimeter (ppmm) for the width."""
-
-        # If the user specifies their monitor size through the Preferences
-        # dialog, then use that size for computing the ppmm.
-        if not self.compute_monitor_size:
-            return self._compute_ppmm()
-
-        if self._monitor and (mm := self._monitor.get_width_mm()):
-            return self.monitor_width / mm
-        else:
-            # Some environments do not provide the monitor size (mm == 0).
+        if not self.current_monitor.width_ppmm:
+            # Some environments do not provide the monitor size.
             # In that case, use a default monitor size of 24 inches, and
-            # warn the the user
+            # warn the the user.
+            logger.debug("width_ppmm == 0")
+            self.current_monitor.set_diagonal_inch(24.0)
+            self.current_monitor.set_compute(False)
             self.warning()
-            return self._compute_ppmm()
+        return self.current_monitor.width_ppmm
 
     @property
     def ppmm_height(self) -> float:
         """Return the number of pixels per millimeter (ppmm) for the height."""
-
-        # If the user specifies their monitor size through the Preferences
-        # dialog, then use that size for computing the ppmm.
-        if not self.compute_monitor_size:
-            return self._compute_ppmm()
-
-        if self._monitor and (mm := self._monitor.get_height_mm()):
-            return self.monitor_height / mm
-        else:
-            # Some environments do not provide the monitor size (mm == 0).
+        if not self.current_monitor.height_ppmm:
+            # Some environments do not provide the monitor size.
             # In that case, use a default monitor size of 24 inches, and
-            # warn the the user
+            # warn the the user.
+            logger.debug("height_ppmm == 0")
+            self.current_monitor.set_diagonal_inch(24.0)
+            self.current_monitor.set_compute(False)
             self.warning()
-            return self._compute_ppmm()
+        return self.current_monitor.height_ppmm
+
+    @property
+    def monitor_width(self) -> int:
+        """Return the display width in pixels."""
+        return self.current_monitor.width_px
+
+    @property
+    def monitor_height(self) -> int:
+        """Return the display height in pixels."""
+        return self.current_monitor.height_px